create wrapper component for AddToDashboardButton
