+++ public/app/features/explore/AddToDashboard/index.tsx
@@ -1,24 +1,49 @@
-import React, { useState } from 'react';
-import { ToolbarButton } from '@grafana/ui';
-import { AddToDashboardModal } from './AddToDashboardModal';
-import { DataQuery } from '@grafana/data';
+import React from 'react';
+import { DataFrame, DataQuery } from '@grafana/data';
+import { ExploreId, StoreState } from 'app/types';
+import { useSelector } from 'react-redux';
+import { getExploreItemSelector } from '../state/selectors';
+import { AddToDashboardButton } from './AddToDashboardButton';
 
+const isActive = (query: DataQuery) => !query.hide;
+const hasRefId = (refId: DataFrame['refId']) => (frame: DataFrame) => frame.refId === refId;
+
+const getMainVisualization = (
+  queries: DataQuery[],
+  graphFrames?: DataFrame[],
+  logsFrames?: DataFrame[],
+  nodeGraphFrames?: DataFrame[]
+) => {
+  for (const { refId } of queries.filter(isActive)) {
+    // traceview is not supported in dashboards, skipping it for now.
+    const hasQueryRefId = hasRefId(refId);
+    if (graphFrames?.some(hasQueryRefId)) {
+      return 'timeseries';
+    }
+    if (logsFrames?.some(hasQueryRefId)) {
+      return 'logs';
+    }
+    if (nodeGraphFrames?.some(hasQueryRefId)) {
+      return 'nodeGraph';
+    }
+  }
+
+  // falling back to table
+  return 'table';
+};
 interface Props {
-  queries: DataQuery[];
-  visualization: string;
+  exploreId: ExploreId;
 }
-export const AddToDashboardButton = ({ queries, visualization }: Props) => {
-  const [isOpen, setIsOpen] = useState(false);
 
-  return (
-    <>
-      <ToolbarButton icon="apps" onClick={() => setIsOpen(true)} aria-label="Add to Dashboard">
-        Add to Dashboard
-      </ToolbarButton>
+export const AddToDashboard = ({ exploreId }: Props) => {
+  const selectExploreItem = getExploreItemSelector(exploreId);
 
-      {isOpen && (
-        <AddToDashboardModal onClose={() => setIsOpen(false)} queries={queries} visualization={visualization} />
-      )}
-    </>
-  );
+  const { queries, mainVisualization } = useSelector((state: StoreState) => {
+    const queries = selectExploreItem(state)?.queries || [];
+    const { graphFrames, logsFrames, nodeGraphFrames } = selectExploreItem(state)?.queryResponse || {};
+
+    return { queries, mainVisualization: getMainVisualization(queries, graphFrames, logsFrames, nodeGraphFrames) };
+  });
+
+  return <AddToDashboardButton queries={queries} visualization={mainVisualization} />;
 };
