+++ public/app/features/explore/ExploreToolbar.tsx
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import { ExploreId } from 'app/types/explore';
 import { PageToolbar, SetInterval, ToolbarButton, ToolbarButtonRow } from '@grafana/ui';
-import { DataFrame, DataQuery, DataSourceInstanceSettings, RawTimeRange } from '@grafana/data';
+import { DataSourceInstanceSettings, RawTimeRange } from '@grafana/data';
 import { DataSourcePicker } from '@grafana/runtime';
 import { StoreState } from 'app/types/store';
 import { createAndCopyShortLink } from 'app/core/utils/shortLinks';
@@ -18,7 +18,7 @@ import { LiveTailControls } from './useLiveTailControls';
 import { cancelQueries, runQueries } from './state/query';
 import { isSplit } from './state/selectors';
 import { DashNavButton } from '../dashboard/components/DashNav/DashNavButton';
-import { AddToDashboardButton } from './AddToDashboard';
+import { AddToDashboard } from './AddToDashboard';
 
 interface OwnProps {
   exploreId: ExploreId;
@@ -71,8 +71,6 @@ class UnConnectedExploreToolbar extends PureComponent<Props> {
       containerWidth,
       onChangeTimeZone,
       onChangeFiscalYearStartMonth,
-      queries,
-      mainVisualization,
     } = this.props;
 
     const showSmallDataSourcePicker = (splitted ? containerWidth < 700 : containerWidth < 800) || false;
@@ -130,7 +128,7 @@ class UnConnectedExploreToolbar extends PureComponent<Props> {
             />
           )}
 
-          <AddToDashboardButton queries={queries} visualization={mainVisualization} />
+          <AddToDashboard exploreId={exploreId} />
 
           <RunButton
             refreshInterval={refreshInterval}
@@ -165,32 +163,6 @@ class UnConnectedExploreToolbar extends PureComponent<Props> {
   }
 }
 
-const isActive = (query: DataQuery) => !query.hide;
-const hasRefId = (refId: DataFrame['refId']) => (frame: DataFrame) => frame.refId === refId;
-
-const getMainVisualization = (
-  queries: DataQuery[],
-  graphFrames: DataFrame[],
-  logsFrames: DataFrame[],
-  nodeGraphFrames: DataFrame[]
-) => {
-  for (const { refId } of queries.filter(isActive)) {
-    // traceview is not supported in dashboards, skipping it for now.
-    const hasQueryRefId = hasRefId(refId);
-    if (graphFrames?.some(hasQueryRefId)) {
-      return 'timeseries';
-    }
-    if (logsFrames?.some(hasQueryRefId)) {
-      return 'logs';
-    }
-    if (nodeGraphFrames?.some(hasQueryRefId)) {
-      return 'nodeGraph';
-    }
-  }
-
-  return 'table';
-};
-
 const mapStateToProps = (state: StoreState, { exploreId }: OwnProps) => {
   const { syncedTimes } = state.explore;
   const exploreItem = state.explore[exploreId]!;
@@ -204,7 +176,6 @@ const mapStateToProps = (state: StoreState, { exploreId }: OwnProps) => {
     isPaused,
     containerWidth,
     queries,
-    queryResponse: { graphFrames, logsFrames, nodeGraphFrames },
   } = exploreItem;
 
   const hasLiveOption = !!datasourceInstance?.meta?.streaming;
@@ -224,7 +195,6 @@ const mapStateToProps = (state: StoreState, { exploreId }: OwnProps) => {
     syncedTimes,
     containerWidth,
     queries,
-    mainVisualization: getMainVisualization(queries, graphFrames, logsFrames, nodeGraphFrames),
   };
 };
